move admin routes to root
