+++ backend/functions/src/student/routes.ts
@@ -9,9 +9,6 @@ import {
   getAllStudents,
   removeStudent,
   updateStudentNotes,
-  addAllowedUser,
-  removeAllowedUser,
-  getAllAllowedUsers,
 } from './functions'
 
 router.get('/', (_, res) => {
@@ -64,37 +61,6 @@ router.delete('/', checkAuth, (req, res) => {
     .catch((err) => {
       console.log(err)
       res.status(400).send({ success: false, err: err.message })
-    })
-})
-
-router.post('/admin', (req, res) => {
-  const { name, email } = req.body
-  addAllowedUser(name, email)
-    .then((data) => res.status(200).send({ success: true, data }))
-    .catch((err) => {
-      console.log(err)
-      res.status(400).send({ success: false, err: err.message })
-    })
-})
-
-router.delete('/admin', (req, res) => {
-  const { email } = req.body
-  removeAllowedUser(email)
-    .then((data) => res.status(200).send({ success: true, data }))
-    .catch((err) => {
-      console.log(err)
-      res.status(400).send({ success: false, err: err.message })
-    })
-})
-
-router.get('/admin', (_, res) => {
-  getAllAllowedUsers()
-    .then((data) => res.status(200).send(data))
-    .catch((err) => {
-      logger.error(
-        `Unexpected error getting all administrators: ${err.message}`
-      )
-      res.status(500).send({ message: err.message })
     })
 })
